Prettify overview page
Overall make up of the overview page with using shadcn card
component.

The delta with open banking account balance is now shown
as an icon which is more concise, when there is no delta an icon
is shown which is more accesible as it is not relying only on the
color.